moved NullOperationLogger to logger module of tracing fixed issue which appears when odl values reference changed due to refinement, therefore added extra log_old_value and log_new_value methods, so that the values are copied when retrieved
